seperate client and server components so that we face no "fs" issues when integrating database
