+++ app/dashboard/types.ts
@@ -9,21 +9,23 @@ export interface Stats {
   weight: number;
 }
 
-export interface Exercise {
-  name: string;
-  reps: string;
-  sets: number;
-  result: string;
-}
-
 export interface Milestone {
   progress: number;
   unit: string;
 }
 
+export interface Exercise {
+  category: string;
+  muscle_group: string;
+  workout: string;
+  equipment: string;
+  sets: number;
+  reps: number;
+  weight: number;
+}
+
 export interface DashboardData {
   user: User;
   stats: Stats;
-  exercises: Exercise[];
   milestones: Record<string, Milestone>;
 }
